+++ settings.php
@@ -113,6 +113,23 @@ if ($hassiteconfig) {
                         get_string('setting_removenodestechnicalhint', 'local_boostnavigation', null, true).'<br />'.
                         get_string('setting_removemycoursesnodeperformancehint', 'local_boostnavigation', null, true),
                 0));
+
+        // Add toggle nodes heading.
+        $page->add(new admin_setting_heading('local_boostnavigation/togglenodesheading',
+                get_string('setting_togglenodesheading', 'local_boostnavigation', null, true),
+                ''));
+
+        // Create my coureses node toggle widget.
+        $page->add(new admin_setting_configcheckbox('local_boostnavigation/togglenodemycourses',
+                get_string('setting_togglenodemycourses', 'local_boostnavigation', null, true),
+                get_string('setting_togglenodemycourses_desc', 'local_boostnavigation', null, true),
+                0));
+
+        // Create course home node toggle widget.
+        $page->add(new admin_setting_configcheckbox('local_boostnavigation/togglenodecoursesections',
+                get_string('setting_togglenodecoursesections', 'local_boostnavigation', null, true),
+                get_string('setting_togglenodecoursesections_desc', 'local_boostnavigation', null, true),
+                0));
     }